+++ packages/server/test/_mocks/notifier/events/index.ts
@@ -1,1 +1,2 @@
 export * from './forum'
+export * from './proposal'